test(cli): snapshot cdk list's user-facing output
Rework the `cdk list` tests to assert on what the user actually sees by
recording the command's full message stream (IoHostRecorder) as committed NDJSON
snapshots, rather than inspecting the raw CDK_TOOLKIT_I2901 payload. The tests
are reframed around user features: listing selected stacks, showing dependency
order (flat chains, nested-by-display-name, and cross-stack), `--long` details,
`--json` output, and failing on circular dependencies.

This exercises the real notify path so the command's output listeners run, which
demonstrates the effective-output recording added in this PR: the `--json`
snapshot shows the synthesis-time line marked `dropped`, and the listing
snapshots show the rendered text rather than the toolkit-lib payload.

Renames list-stacks.test.ts to list.test.ts; the formatStackList unit tests are
unchanged.
