+++ packages/aws-cdk/test/commands/__io_snapshots__/list/cdk_list_lists_the_selected_stacks.ndjson
@@ -1,0 +1,5 @@
+{"seq":0,"type":"notify","action":"list","level":"trace","code":"CDK_TOOLKIT_I1001","message":"Starting Synthesis ..."}
+{"seq":1,"type":"notify","action":"deploy","level":"trace","code":"CDK_CLI_I1000","message":"Starting Synthesis ..."}
+{"seq":2,"type":"notify","action":"deploy","level":"trace","code":"CDK_CLI_I1001","message":"\n✨  Synthesis time: <DURATION>\n"}
+{"seq":3,"type":"notify","action":"list","level":"info","code":"CDK_TOOLKIT_I1000","message":"✨  Synthesis time: <DURATION>"}
+{"seq":4,"type":"notify","action":"list","level":"result","code":"CDK_TOOLKIT_I2901","message":"Test-Stack-A\nTest-Stack-B"}
